Rename nostr module for API
